+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -12,7 +12,9 @@
 
 ## Version 2.6.0 (dev)
 
+* Added `Reference::isEmpty`
 * Empty strings are now detected as invalid in the `SiteLink` constructor
+* Empty References are now ignored when added to `ReferenceList`
 * The `ReferenceList` constructor now throws an `InvalidArgumentException` when getting a non-iterable input
 * The `SnakList` constructor now throws an `InvalidArgumentException` when getting a non-iterable input
 * The `AliasGroup::equals` and `Term::equals` methods no longer incorrectly return true for fallback objects